+++ deps/cli.ts
@@ -2,4 +2,4 @@
 
 export * from "./common.ts";
 
-export * as cliffy_cmd from "https://deno.land/x/cliffy@v1.0.0-rc.3/command/mod.ts";
+export * as cliffy_cmd from "https://deno.land/x/cliffy@v1.0.0-rc.4/command/mod.ts";
